Fixed TTS (re)start handling
The issue was, that when the cookie changed during build (by 'make
rel'), running './tts restart' can't work afterwards, since the running
TTS still expects the old cookie.
So now we stop the tts before build, and then start it afterwards
again (therefore always knowing the correct cookie).

A better solution might be to only regenerate a cookie (and restart the
TTS) when the build changes anything, but since all of this will need to
be updates once @lukasburgey is done with the packages (rpm, deb) this
is good for now.